+++ exporter/opentelemetry-exporter-opencensus/src/opentelemetry/exporter/opencensus/trace_exporter/__init__.py
@@ -4,7 +4,7 @@
 """OpenCensus Span Exporter."""
 
 import logging
-from typing import Sequence
+from collections.abc import Sequence
 
 import grpc
 from opencensus.proto.agent.trace.v1 import (