+++ opentelemetry-api/src/opentelemetry/context/__init__.py
@@ -4,7 +4,6 @@
 from __future__ import annotations
 
 import logging
-import typing
 from contextvars import Token
 from os import environ
 from uuid import uuid4
@@ -71,7 +70,7 @@ def create_key(keyname: str) -> str:
     return keyname + "-" + str(uuid4())
 
 
-def get_value(key: str, context: typing.Optional[Context] = None) -> object:
+def get_value(key: str, context: Context | None = None) -> object:
     """To access the local state of a concern, the RuntimeContext API
     provides a function which takes a context and a key as input,
     and returns a value.
@@ -87,7 +86,7 @@ def get_value(key: str, context: typing.Optional[Context] = None) -> object:
 
 
 def set_value(
-    key: str, value: object, context: typing.Optional[Context] = None
+    key: str, value: object, context: Context | None = None
 ) -> Context:
     """To record the local state of a cross-cutting concern, the
     RuntimeContext API provides a function which takes a context, a