+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/_internal/exemplar/exemplar.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import dataclasses
-from typing import Optional, Union
 
 from opentelemetry.util.types import Attributes
 
@@ -28,7 +27,7 @@ class Exemplar:
     """
 
     filtered_attributes: Attributes
-    value: Union[int, float]
+    value: int | float
     time_unix_nano: int
-    span_id: Optional[int] = None
-    trace_id: Optional[int] = None
+    span_id: int | None = None
+    trace_id: int | None = None