+++ opentelemetry-sdk/src/opentelemetry/sdk/util/__init__.py
@@ -6,7 +6,6 @@ import datetime
 import threading
 from collections import deque
 from collections.abc import MutableMapping, Sequence
-from typing import Optional
 
 from typing_extensions import deprecated
 
@@ -40,7 +39,7 @@ class BoundedList(Sequence):
     not enough room.
     """
 
-    def __init__(self, maxlen: Optional[int]):
+    def __init__(self, maxlen: int | None):
         self.dropped = 0
         self._dq = deque(maxlen=maxlen)  # type: deque
         self._lock = threading.Lock()
@@ -99,7 +98,7 @@ class BoundedDict(MutableMapping):
     added.
     """
 
-    def __init__(self, maxlen: Optional[int]):
+    def __init__(self, maxlen: int | None):
         if maxlen is not None:
             if not isinstance(maxlen, int):
                 raise ValueError