+++ tests/opentelemetry-test-utils/src/opentelemetry/test/mock_textmap.py
@@ -1,7 +1,6 @@
 # Copyright The OpenTelemetry Authors
 # SPDX-License-Identifier: Apache-2.0
 
-import typing
 
 from opentelemetry import trace
 from opentelemetry.context import Context
@@ -25,7 +24,7 @@ class NOOPTextMapPropagator(TextMapPropagator):
     def extract(
         self,
         carrier: CarrierT,
-        context: typing.Optional[Context] = None,
+        context: Context | None = None,
         getter: Getter = default_getter,
     ) -> Context:
         return Context()
@@ -33,7 +32,7 @@ class NOOPTextMapPropagator(TextMapPropagator):
     def inject(
         self,
         carrier: CarrierT,
-        context: typing.Optional[Context] = None,
+        context: Context | None = None,
         setter: Setter = default_setter,
     ) -> None:
         return None
@@ -52,7 +51,7 @@ class MockTextMapPropagator(TextMapPropagator):
     def extract(
         self,
         carrier: CarrierT,
-        context: typing.Optional[Context] = None,
+        context: Context | None = None,
         getter: Getter = default_getter,
     ) -> Context:
         if context is None:
@@ -77,7 +76,7 @@ class MockTextMapPropagator(TextMapPropagator):
     def inject(
         self,
         carrier: CarrierT,
-        context: typing.Optional[Context] = None,
+        context: Context | None = None,
         setter: Setter = default_setter,
     ) -> None:
         span = trace.get_current_span(context)